% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break-point.R
\name{read_rdp_breakpoint}
\alias{read_rdp_breakpoint}
\title{read_rdp_breakpoint}
\usage{
read_rdp_breakpoint(file)
}
\arguments{
\item{file}{break point file}
}
\value{
A 'RDP_breakPoint' object
}
\description{
Parse break point output from RDP5
}
\details{
This function read the exported csv files from RDP5 breakpoint plot
}
\author{
Guangchuang Yu
}
